replace tokenizer (#205)
